% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-hbr.R
\docType{methods}
\name{inspect_window_PCA,hbr-method}
\alias{inspect_window_PCA}
\alias{inspect_window_PCA,hbr-method}
\alias{inspect_window_PCA.hbr}
\title{PCA of haplotype-block marker for a window}
\usage{
\S4method{inspect_window_PCA}{hbr}(object, window)
}
\arguments{
\item{object}{object of class \code{\link{hbr}} containing the results of 
reconstructed haplotype-blocks}

\item{window}{number of the window, which should be inspected}
}
\description{
Plots all filtered SNPs for a given window in terms of their the 
first two principal components of the respective time series
data.
}
\details{
Performs principal component analysis (PCA) on the time series data 
of the filtered SNPs and the selected time points (filtered and selected 
through \code{use.libs} in \code{\link{initialize_SNP_time_series}})
in a given window. SNPs are shown in terms of 
the first two principal components and colored according to the
haplotype-block they were assigned to. Black colored points were 
not assigned to any cluster.
}
\author{
Susanne U. Franssen
}
\seealso{
\code{\link{hbr}} \code{\link{ex_dat}} \code{\link{summary.hbr}} \code{\link{plot.hbr}} 
\code{\link{plot_cluster_trajectories}} \code{\link{plot_marker_trajectories}}
\code{\link{map}} \code{\link{rev_map}} \code{\link{markers}} \code{\link{plot_hbr_freq}}
\code{\link{inspect_window}} \code{\link{inspect_window_avLink}}
\code{\link{inspect_window_dbScan}} \code{\link{number_hbr}}
}

