% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-hbr.R
\docType{methods}
\name{markers,hbr-method}
\alias{markers}
\alias{markers,hbr-method}
\alias{markers.hbr}
\title{Markers of the specified reconstructed haplotype-block}
\usage{
\S4method{markers}{hbr}(object, hbr_id)
}
\arguments{
\item{object}{object of class \code{\link{hbr}} containing the results of reconstructed 
haplotype-blocks}

\item{hbr_id}{index of the haplotype-block of interest}
}
\value{
numeric vector with marker positions of the specified reconstructed haplotype-block
}
\description{
Returns the genomic positions of the markers for the specified reconstructed haplotype-block
}
\author{
Susanne U. Franssen
}
\seealso{
\code{\link{hbr}} \code{\link{ex_dat}} \code{\link{summary.hbr}} \code{\link{plot.hbr}} 
\code{\link{plot_cluster_trajectories}} \code{\link{plot_marker_trajectories}}
\code{\link{map}} \code{\link{rev_map}} \code{\link{plot_hbr_freq}}
\code{\link{inspect_window}} \code{\link{inspect_window_PCA}} \code{\link{inspect_window_avLink}}
\code{\link{inspect_window_dbScan}} \code{\link{number_hbr}}
}

