% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-hbr.R
\docType{methods}
\name{number_hbr,hbr-method}
\alias{number_hbr}
\alias{number_hbr,hbr-method}
\alias{number_hbr.hbr}
\title{The number of reconstructed haplotype-block}
\usage{
\S4method{number_hbr}{hbr}(object)
}
\arguments{
\item{object}{object of class \code{\link{hbr}} containing the results of reconstructed 
haplotype-blocks}
}
\value{
returns the number of reconstructed haplotype-blocks
}
\description{
Returns the number of reconstructed haplotype-blocks
}
\author{
Susanne U. Franssen
}
\seealso{
\code{\link{hbr}} \code{\link{ex_dat}} \code{\link{summary.hbr}} \code{\link{plot.hbr}} 
\code{\link{plot_cluster_trajectories}} \code{\link{plot_marker_trajectories}}
\code{\link{map}} \code{\link{rev_map}} \code{\link{markers}} \code{\link{plot_hbr_freq}}
\code{\link{inspect_window}} \code{\link{inspect_window_PCA}} \code{\link{inspect_window_avLink}}
\code{\link{inspect_window_dbScan}}
}

