% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-hbr.R
\docType{methods}
\name{rev_map,hbr-method}
\alias{rev_map}
\alias{rev_map,hbr-method}
\alias{rev_map.hbr}
\title{Map from windows to reconstructed haplotype-blocks}
\usage{
\S4method{rev_map}{hbr}(object)
}
\arguments{
\item{object}{object of class \code{\link{hbr}} containing the results of reconstructed 
haplotype-blocks}
}
\value{
list with elements corresponding to all windows that contain a reconstructed
haplotype-block segment. Each element contains a vector of haplotype-block indices
that overlap the respective window.
}
\description{
A method returning information which reconstructed haplotype-blocks are present in 
which windows.
}
\details{
The method operates on \code{\link{hbr}} objects and returns a list summarizing
the mapping information from windows to reconstructed haplotype-blocks.
}
\author{
Susanne U. Franssen
}
\seealso{
\code{\link{hbr}} \code{\link{ex_dat}} \code{\link{summary.hbr}} \code{\link{plot.hbr}} 
\code{\link{plot_cluster_trajectories}} \code{\link{plot_marker_trajectories}}
\code{\link{map}} \code{\link{markers}} \code{\link{plot_hbr_freq}}
\code{\link{inspect_window}} \code{\link{inspect_window_PCA}} \code{\link{inspect_window_avLink}}
\code{\link{inspect_window_dbScan}} \code{\link{number_hbr}}
}

