% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-hbr.R
\docType{methods}
\name{show,hbr-method}
\alias{show}
\alias{show,hbr-method}
\alias{show.hbr}
\title{Display of a hbr object}
\usage{
\S4method{show}{hbr}(object)
}
\arguments{
\item{object}{object of the class hbr data}
}
\description{
This function displays the summarized features of a hbr object
}
\details{
This function displays the summarized features of a hbr object
}
\author{
Susanne U. Franssen
}
\references{
Franssen et al. 2016, \href{http://mbe.oxfordjournals.org/content/early/2016/10/03/molbev.msw210.abstract}{Reconstruction of haplotype-blocks
selected during experimental evolution}, \href{http://mbe.oxfordjournals.org/}{MBE}
}
\seealso{
\code{\link{summary.hbr}} \code{\link{ex_dat}} \code{\link{plot.hbr}} 
\code{\link{plot_cluster_trajectories}} \code{\link{plot_marker_trajectories}}
\code{\link{map}} \code{\link{rev_map}} \code{\link{markers}} \code{\link{plot_hbr_freq}}
\code{\link{inspect_window}} \code{\link{inspect_window_PCA}} \code{\link{inspect_window_avLink}}
\code{\link{inspect_window_dbScan}} \code{\link{number_hbr}}
}

