% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNP_time_series.R
\docType{methods}
\name{show,SNP_time_series-method}
\alias{show,SNP_time_series-method}
\alias{show.SNP_time_series}
\title{Display of a SNP_time_series object}
\usage{
\S4method{show}{SNP_time_series}(object)
}
\arguments{
\item{object}{object of the class SNP_time_series data}
}
\description{
This function displays the summarized features of a SNP_time_series object
}
\details{
This function displays the summarized features of a SNP_time_series object
}
\author{
Susanne U. Franssen
}
\references{
Franssen et al. 2016, \href{http://mbe.oxfordjournals.org/content/early/2016/10/03/molbev.msw210.abstract}{Reconstruction of haplotype-blocks
selected during experimental evolution}, \href{http://mbe.oxfordjournals.org/}{MBE}
}
\seealso{
\code{\link{ex_dat}} \code{\link{initialize_SNP_time_series}} \code{\link{SNP_time_series}}
}

